+++ src/system/STM32F10x_StdPeriph_Driver/src/stm32f10x_i2c.c
@@ -607,7 +607,7 @@ void I2C_Send7bitAddress(I2C_TypeDef* I2Cx, uint8_t Address, uint8_t I2C_Directi
   if (I2C_Direction != I2C_Direction_Transmitter)
   {
     /* Set the address bit0 for read */
-    Address |= OAR1_ADD0_Set;
+    Address |= (uint8_t)OAR1_ADD0_Set;
   }
   else
   {
